Create LSTM_Chapel_GPT4.chpl
This code defines a simple LSTM unit with methods to initialize weights and forward propagation. Note that this implementation does not include training functionality, so the LSTM unit won't learn from data. To train the LSTM unit, you will need to implement a backpropagation through time algorithm and a training loop.